requirements.txt
pylint==2.12.2
pytest==6.2.5
black==21.12b0
torch==1.10.1
torchvision==0.11.2
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.1
pytest-cov==3.0.0
pytorch-lightning==1.5.7
scikit-learn==1.0.2
matplotlib==3.5.1
pandas==1.3.5
wandb==0.12.9
omegaconf==2.1.1
hydra-core==1.1.1
seaborn==0.11.2

